Refactor + cancel premium limitations
