+++ app/realms/__init__.py
@@ -7,3 +7,4 @@ from app.realms.badges import routes
 from app.realms.users import routes
 from app.realms.leaderboards import routes
 from app.realms.rewards import routes
+from app.realms.premium import routes
