+++ app/realms/routes.py
@@ -1,4 +1,4 @@
-from app import stripe, db
+from app import db
 from flask import render_template, flash, redirect, url_for
 from flask_login import current_user, login_user, logout_user, login_required
 from app.models import Admin, Realm, User, Badge
@@ -125,58 +125,6 @@ def delete_realm(id):
 
     flash('Realm not deleted because passwords didn\'t match.')
     return redirect(url_for('realms.show_realm', id=id))
-
-
-
-@bp.route('/realms/payment', methods=['POST'])
-@login_required
-def payment():
-    customer = stripe.Customer.list(email=request.form['stripeEmail'])
-    
-    if customer.data:
-        customer = customer.data[0]
-    else:
-        customer = stripe.Customer.create(
-            email=request.form['stripeEmail'],
-            source=request.form['stripeToken']
-        )
-
-    subscription = stripe.Subscription.create(
-        customer=customer.id,
-        items=[{"plan": "plan_H7quRNGUfctwNA"}],
-    )
-    
-    admin = Admin.query.get_or_404(current_user.get_id())
-    admin.premium = 1
-    admin.subscription_key = subscription.id
-
-    db.session.add(admin)
-    db.session.commit()
-
-    flash('Congratulations, you now have a Premium account!')
-    return redirect(url_for('realms.realms'))
-
-
-@bp.route('/realms/premium')
-@login_required
-def premium():
-    admin = Admin.query.get_or_404(current_user.get_id())
-
-    return render_template('realms/premium.html', admin=admin, key=stripe.publishable_key)
-
-@bp.route('/realms/cancel')
-@login_required
-def cancel():
-    admin = Admin.query.get_or_404(current_user.get_id())
-    stripe.Subscription.delete(admin.subscription_key)
-
-    admin.premium = 0
-    admin.subscription_key = None
-    db.session.add(admin)
-    db.session.commit()
-
-    flash('You\'ve cancelled your Premium account!')
-    return redirect(url_for('realms.realms'))
 
 
 @bp.route('/realms/settings')
